[Silver II] Title: 잃어버린 괄호, Time: 0 ms, Memory: 2028 KB -BaekjoonHub

+++ 백준/Silver/1541. 잃어버린 괄호/잃어버린 괄호.cc
@@ -33,7 +33,7 @@ int main() {
 		part = part.substr(pos + 1);
 	}
 
-	for (int i = 1; i < (int)str.size(); i++) {
+	for (int i = 1; i < str.size(); i++) {
 		part = str[i];
 		int temp = 0;
